e2e/nats-loki-bridge/requirements.txt
nats-py>=2.7.0
